fix: enforce env strictness and fix test DB isolation
docker-compose.yml
- Apply ${POSTGRES_PASSWORD:?...} to fail fast on missing env var
- Add restart: on-failure to skinmate-graph-seeder

db/initdb/05-traverse_cache.sql
- Change user_id type INT -> BIGINT (align with memories table)
- Unify RLS policy casting from ::integer to ::bigint

scripts/populate_graph.py
- Remove POSTGRES_PASSWORD hardcoded fallback; exit(1) if unset
- Convert timeout fallback path to exit(1) to trigger Docker restart
- Move import time to module top level

tests/conftest.py
- Add _ensure_test_db(): create skinmate_test from TEMPLATE skinmate
  with pg_terminate_backend for concurrent connection handling,
  then TRUNCATE user-data tables on fresh creation
- Set DATABASE_URL to app user (skinmate_app) with db name replaced,
  so RLS policies are enforced during tests
- Centralize db_conn fixture (superuser) and clean_db_fixtures (full DELETE)

tests/graph/test_traverse.py
- Remove duplicate fixture_db_conn (now in conftest.py)
- Wrap memories INSERT/UPDATE with db.user_scope for RLS compliance

tests/knowledge/test_hard_filter.py
tests/retrieval/test_retrieve.py
- Remove duplicate fixture_db_conn (now in conftest.py)

All 103 tests pass (91 unit + 12 contract).

+++ db/initdb/05-traverse_cache.sql
@@ -6,7 +6,7 @@ SET search_path = public;
 
 -- 1. 캐시 테이블 생성
 CREATE TABLE IF NOT EXISTS public.traverse_cache (
-    user_id INT NOT NULL,
+    user_id BIGINT NOT NULL,
     season VARCHAR(50) NOT NULL,
     paths_json JSONB NOT NULL,
     created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
@@ -22,8 +22,8 @@ ALTER TABLE public.traverse_cache FORCE  ROW LEVEL SECURITY;
 DROP POLICY IF EXISTS traverse_cache_isolation_policy ON public.traverse_cache;
 CREATE POLICY traverse_cache_isolation_policy ON public.traverse_cache
     FOR ALL
-    USING (user_id = NULLIF(current_setting('app.current_user_id', true), '')::integer)
-    WITH CHECK (user_id = NULLIF(current_setting('app.current_user_id', true), '')::integer);
+    USING (user_id = NULLIF(current_setting('app.current_user_id', true), '')::bigint)
+    WITH CHECK (user_id = NULLIF(current_setting('app.current_user_id', true), '')::bigint);
 
 -- 4. skinmate_app 역할에 권한 부여
 GRANT SELECT, INSERT, UPDATE, DELETE ON public.traverse_cache TO skinmate_app;
